feat(storage-client): benchmark test
